+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add delete() method to Account
 - Add move() method to Account
 - Add reorder() method to Account
+- Add merge() method to Account
 - Add proper testing for ToshlException
 - Improve string representation of ToshlException
 - refactoring: each entity is in a separate module